+++ testutils_robolectric/src/main/java/com/google/android/exoplayer2/testutil/MediaSourceTestRunner.java
@@ -47,6 +47,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
 
 /** A runner for {@link MediaSource} tests. */
 public class MediaSourceTestRunner {
@@ -62,6 +63,9 @@ public class MediaSourceTestRunner {
 
   private final LinkedBlockingDeque<Timeline> timelines;
   private final CopyOnWriteArrayList<Pair<Integer, MediaPeriodId>> completedLoads;
+  private final AtomicReference<MediaPeriodId> lastCreatedMediaPeriod;
+  private final AtomicReference<MediaPeriodId> lastReleasedMediaPeriod;
+
   private Timeline timeline;
 
   /**
@@ -79,6 +83,8 @@ public class MediaSourceTestRunner {
     mediaSourceListener = new MediaSourceListener();
     timelines = new LinkedBlockingDeque<>();
     completedLoads = new CopyOnWriteArrayList<>();
+    lastCreatedMediaPeriod = new AtomicReference<>();
+    lastReleasedMediaPeriod = new AtomicReference<>();
     mediaSource.addEventListener(playbackHandler, mediaSourceListener);
   }
 
@@ -282,16 +288,20 @@ public class MediaSourceTestRunner {
   private void assertPrepareAndReleasePeriod(MediaPeriodId mediaPeriodId)
       throws InterruptedException {
     MediaPeriod mediaPeriod = createPeriod(mediaPeriodId);
+    assertThat(lastCreatedMediaPeriod.getAndSet(/* newValue= */ null)).isEqualTo(mediaPeriodId);
     CountDownLatch preparedCondition = preparePeriod(mediaPeriod, 0);
     assertThat(preparedCondition.await(TIMEOUT_MS, TimeUnit.MILLISECONDS)).isTrue();
     // MediaSource is supposed to support multiple calls to createPeriod with the same id without an
     // intervening call to releasePeriod.
     MediaPeriod secondMediaPeriod = createPeriod(mediaPeriodId);
+    assertThat(lastCreatedMediaPeriod.getAndSet(/* newValue= */ null)).isEqualTo(mediaPeriodId);
     CountDownLatch secondPreparedCondition = preparePeriod(secondMediaPeriod, 0);
     assertThat(secondPreparedCondition.await(TIMEOUT_MS, TimeUnit.MILLISECONDS)).isTrue();
     // Release the periods.
     releasePeriod(mediaPeriod);
+    assertThat(lastReleasedMediaPeriod.getAndSet(/* newValue= */ null)).isEqualTo(mediaPeriodId);
     releasePeriod(secondMediaPeriod);
+    assertThat(lastReleasedMediaPeriod.getAndSet(/* newValue= */ null)).isEqualTo(mediaPeriodId);
   }
 
   /**
@@ -355,6 +365,18 @@ public class MediaSourceTestRunner {
     // MediaSourceEventListener methods.
 
     @Override
+    public void onMediaPeriodCreated(int windowIndex, MediaPeriodId mediaPeriodId) {
+      Assertions.checkState(Looper.myLooper() == playbackThread.getLooper());
+      lastCreatedMediaPeriod.set(mediaPeriodId);
+    }
+
+    @Override
+    public void onMediaPeriodReleased(int windowIndex, MediaPeriodId mediaPeriodId) {
+      Assertions.checkState(Looper.myLooper() == playbackThread.getLooper());
+      lastReleasedMediaPeriod.set(mediaPeriodId);
+    }
+
+    @Override
     public void onLoadStarted(
         int windowIndex,
         @Nullable MediaPeriodId mediaPeriodId,
@@ -390,6 +412,11 @@ public class MediaSourceTestRunner {
         MediaLoadData mediaLoadData,
         IOException error,
         boolean wasCanceled) {
+      Assertions.checkState(Looper.myLooper() == playbackThread.getLooper());
+    }
+
+    @Override
+    public void onReadingStarted(int windowIndex, MediaPeriodId mediaPeriodId) {
       Assertions.checkState(Looper.myLooper() == playbackThread.getLooper());
     }